修改减速问题bug
修改减速问题bug

+++ 2025广东省大学生游戏开发/Assets/Scripts/StartPos/StartPos.cs
@@ -14,7 +14,7 @@ public class StartPos : MonoBehaviour
             if(player.playerType==playerType)
             {
                 AudioManager.Instance.PlayAudio("Score");
-                GameManager.Instance.ChangeScore(player,player.GetScore());
+                GameManager.Instance.ChangeScore(player,(int)player.GetScore());
                 player.ChangeScore(-player.GetScore());
                 player.GetMovement().ResetSpeed();
             }
